requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.0.0
django-bootstrap4==0.0.8
django-extensions==2.2.1
django-filter==2.1.0
django-fontawesome==1.0
django-mathfilters==0.4.0
django-taggit==1.1.0
Pillow==6.0.0
pytz==2019.1
PyYAML==5.1.1
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
uWSGI==2.0.18
